+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 * Refactor Edit page to support multiple resource types. Add Hub create page. Refs [UILD-694].
 * Persist Hub when save button is clicked. Refs [UILD-699].
 * Add workspace profile settings components. Refs [UILD-696].
+* Preserve search results state. Refs [UILD-702].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -136,6 +137,7 @@
 [UILD-694]:https://folio-org.atlassian.net/browse/UILD-694
 [UILD-699]:https://folio-org.atlassian.net/browse/UILD-699
 [UILD-696]:https://folio-org.atlassian.net/browse/UILD-696
+[UILD-702]:https://folio-org.atlassian.net/browse/UILD-702
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].